데이터베이스 시딩

+++ database/migrations/2019_03_05_063339_create_articles_table.php
@@ -14,8 +14,8 @@ class CreateArticlesTable extends Migration
     public function up()
     {
         Schema::create('articles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->string('title');
             $table->text('content');
             $table->timestamps();
